backend/requirements.txt
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
amqp==5.2.0
antlr4-python3-runtime==4.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
botocore==1.35.36
cachetools==5.5.0
capstone==5.0.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.1.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
fsspec==2024.9.0
gcsfs==2024.9.0.post1
google-api-core==2.21.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
kombu==5.4.2
Markdown==3.7
minio==7.2.9
multidict==6.1.0
oauthlib==3.2.2
pefile==2024.8.26
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3fs==2024.9.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stix2==3.0.1
stix2-patterns==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
volatility3==2.8.0
wcwidth==0.2.13
wrapt==1.16.0
yara-python==4.5.1
yarl==1.14.0
zipp==3.20.2
